+++ src/main/java/sc/fiji/oc3dsd/engine/StarDistTrackMateRunner.java
@@ -17,6 +17,8 @@ import ij.measure.Calibration;
 import ij.measure.ResultsTable;
 import ij.process.ImageProcessor;
 
+import sc.fiji.oc3d.core.label.LabelRenumberer;
+
 import java.io.File;
 import java.util.ArrayList;
 import java.util.HashMap;